requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
flask_cors==4.0.1
jsonschema==4.23.0
python-dotenv==1.0.1
pydantic==2.8.2
pytest==8.3.2
bson==0.5.10